Don't use foscam lib daemon option
It's currently redundant and can cause a thread creating error.

On error flow the thread sleep 30 seconds but the main loop continues
so we eventually create 30 threads.
If the camera is dead for some time we eventualy get an error
that we cannot start new threads as we already have too many.

Currently we don't need the daemon option so we do a new request only
if the first request finished.

also log the error code in the error msg.

Signed-off-by: Roi Dayan <roi.dayan@gmail.com>